requirements_ginger.txt
absl-py @ file:///tmp/build/80754af9/absl-py_1623861369333/work
actionlib==1.12.1
aiohttp @ file:///tmp/build/80754af9/aiohttp_1637857092750/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
angles==1.9.12
async-timeout @ file:///tmp/build/80754af9/async-timeout_1637851218186/work
asynctest==0.13.0
attrs @ file:///tmp/build/80754af9/attrs_1620827162558/work
base_local_planner==1.16.7
blinker==1.4
bondpy==1.8.5
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1619597386817/work
camera_calibration==1.15.0
camera_calibration_parsers==1.11.13
catkin==0.7.29
catkin-pkg==0.4.24
certifi==2021.10.8
cffi @ file:///tmp/build/80754af9/cffi_1636541929935/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.0.3
controller_manager==0.18.3
controller_manager_msgs==0.18.3
cryptography @ file:///tmp/build/80754af9/cryptography_1633521537906/work
cv_bridge==1.13.0
cycler==0.11.0
diagnostic_analysis==1.9.7
diagnostic_common_diagnostics==1.9.7
diagnostic_updater==1.9.7
distro==1.6.0
docutils==0.18.1
drone_wrapper==1.3.8
dynamic_reconfigure==1.6.3
dynamixel_sdk==3.7.51
fonttools==4.28.3
frozenlist @ file:///tmp/build/80754af9/frozenlist_1637767116045/work
gazebo_plugins==2.8.7
gazebo_ros==2.8.7
gencpp==0.6.5
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.16
gennodejs==2.0.1
genpy==0.6.15
google-auth @ file:///tmp/build/80754af9/google-auth_1626320605116/work
google-auth-oauthlib==0.4.1
gps_common==0.3.1
grpcio @ file:///tmp/build/80754af9/grpcio_1637578017926/work
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
image_geometry==1.13.0
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1638529598726/work
interactive-markers==1.11.5
joint_state_publisher==1.12.15
joint_state_publisher_gui==1.12.15
jsk-recognition-utils==1.2.15
jsk_rviz_plugins==2.1.7
jsk_topic_tools==2.2.11
kdl_parser_py==1.13.1
kiwisolver==1.3.2
laser_geometry==1.6.7
Markdown @ file:///tmp/build/80754af9/markdown_1614363832606/work
matplotlib==3.5.1
mavros==1.7.1
message_filters==1.14.10
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
mpi4py @ file:///tmp/build/80754af9/mpi4py_1594373948151/work
multidict @ file:///tmp/build/80754af9/multidict_1607367738239/work
netifaces==0.11.0
nmea_navsat_driver==0.5.2
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1634106693478/work
oauthlib @ file:///tmp/build/80754af9/oauthlib_1623060228408/work
olefile==0.46
packaging==21.3
Pillow==8.4.0
protobuf==3.17.2
py-trees==0.6.9
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
PyJWT @ file:///tmp/build/80754af9/pyjwt_1619651626416/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1635333100036/work
pyparsing==3.0.6
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
python-dateutil==2.8.2
python_qt_binding==0.4.3
PyYAML==6.0
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
qt_create==0.2.10
requests @ file:///tmp/build/80754af9/requests_1629994808627/work
requests-oauthlib==1.3.0
resource_retriever==1.12.6
rosapi==0.11.13
rosbag==1.14.10
rosboost-cfg==1.14.9
rosbridge_library==0.11.13
rosbridge_server==0.11.13
rosclean==1.14.9
roscreate==1.14.9
rosgraph==1.14.10
roslaunch==1.14.10
roslib==1.14.9
roslint==0.11.2
roslz4==1.14.10
rosmake==1.14.9
rosmaster==1.14.10
rosmsg==1.14.10
rosnode==1.14.10
rosparam==1.14.10
rospkg==1.3.0
rospy==1.14.10
rosserial_arduino==0.8.0
rosserial_client==0.8.0
rosserial_python==0.8.0
rosservice==1.14.10
rostest==1.14.10
rostopic==1.14.10
rosunit==1.14.9
roswtf==1.14.10
rqt-moveit==0.5.10
rqt-reconfigure==0.5.4
rqt-robot-monitor==0.5.13
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_console==0.4.9
rqt_controller_manager==0.18.3
rqt_dep==0.4.9
rqt_drone_teleop==1.3.8
rqt_ez_publisher==0.5.0
rqt_graph==0.4.11
rqt_ground_robot_teleop==1.3.8
rqt_gui==0.5.2
rqt_gui_py==0.5.2
rqt_image_view==0.4.16
rqt_joint_trajectory_controller==0.17.2
rqt_joint_trajectory_plot==0.0.5
rqt_launch==0.4.8
rqt_launchtree==0.2.0
rqt_logger_level==0.4.8
rqt_msg==0.4.8
rqt_multiplot==0.0.10
rqt_nav_view==0.5.7
rqt_play_motion_builder==1.0.2
rqt_plot==0.4.13
rqt_pose_view==0.5.8
rqt_publisher==0.4.8
rqt_py_common==0.5.2
rqt_py_console==0.4.8
rqt_py_trees==0.3.1
rqt_robot_dashboard==0.5.7
rqt_robot_steering==0.5.10
rqt_rotors==2.2.3
rqt_runtime_monitor==0.5.7
rqt_rviz==0.6.0
rqt_service_caller==0.4.8
rqt_shell==0.4.9
rqt_srv==0.4.8
rqt_tf_tree==0.6.0
rqt_top==0.4.8
rqt_topic==0.4.11
rqt_virtual_joy==0.1.2
rqt_web==0.4.8
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
rviz==1.13.17
sensor-msgs==1.12.8
six @ file:///tmp/build/80754af9/six_1623709665295/work
smach==2.0.1
smach_ros==2.0.1
smclib==1.8.5
sound-play==0.3.11
tensorboard @ file:///tmp/build/80754af9/tensorboard_1633093581375/work/tensorboard-2.6.0-py3-none-any.whl
tensorboard-data-server @ file:///tmp/build/80754af9/tensorboard-data-server_1633024562496/work/tensorboard_data_server-0.6.0-py3-none-manylinux2010_x86_64.whl
tensorboard-plugin-wit==1.6.0
tf==1.12.1
tf2_geometry_msgs==0.6.5
tf2_kdl==0.6.5
tf2_py==0.6.5
tf2_ros==0.6.5
tf_conversions==1.12.1
topic_tools==1.14.10
torch==1.7.1
torch-scatter @ file:///home/joeray/Downloads/torch_scatter-2.0.5-cp37-cp37m-linux_x86_64.whl
torchaudio==0.7.0a0+a853dff
torchvision @ file:///tmp/build/80754af9/torchvision_1622219711511/work
turtlebot3_example==1.2.5
turtlebot3_teleop==1.2.5
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
unique_id==1.0.6
urdfdom-py==0.4.5
urllib3==1.26.7
Werkzeug @ file:///tmp/build/80754af9/werkzeug_1635505089296/work
xacro==1.13.10
yarl @ file:///tmp/build/80754af9/yarl_1606939952652/work
zipp @ file:///tmp/build/80754af9/zipp_1633618647012/work
